separate NavDrawer component and the data that initializes the main site nav, put NavDrawer in ol-components and make it reusable
