+++ frontends/ol-components/src/index.ts
@@ -126,6 +126,7 @@ export * from "./components/Chips/ChipLink"
 export * from "./components/EmbedlyCard/EmbedlyCard"
 export * from "./components/FormDialog/FormDialog"
 export * from "./components/RoutedDrawer/RoutedDrawer"
+export * from "./components/NavDrawer/NavDrawer"
 export * from "./components/SearchInput/SearchInput"
 export * from "./components/SimpleMenu/SimpleMenu"
 export * from "./components/SortableList/SortableList"
